+++ tests/testthat/test-fixest.R
@@ -182,23 +182,9 @@ if (require("testthat") &&
   })
 
   test_that("get_statistic", {
-    expect_equal(
-      get_statistic(m1),
-      structure(list(
-        Parameter = "log(dist_km)", Statistic = -13.2102796112222
-      ),
-      class = "data.frame", row.names = c(NA, -1L), statistic = "z-statistic"
-      ),
-      tolerance = 1e-4
-    )
-    expect_equal(
-      get_statistic(m2),
-      structure(list(
-        Parameter = "log(dist_km)", Statistic = -14.0627647683725
-      ),
-      class = "data.frame", row.names = c(NA, -1L), statistic = "z-statistic"
-      ),
-      tolerance = 1e-4
-    )
+    stat <- get_statistic(m1)
+    expect_equal(stat$Statistic, -13.212695, tolerance = 1e-3)
+    stat <- get_statistic(m2)
+    expect_equal(stat$Statistic, -14.065336, tolerance = 1e-3)
   })
 }