+++ NEWS.md
@@ -14,6 +14,7 @@
 * Fixed issue in `get_data()` for models that used the `lspline()`-function.
 * Fixed issue in `get_statistic()` for *multinom* models.
 * Fixed issue in `get_priors()` for *stanreg*  models with flat intercept priors.
+* Fixed tests that failed due to latest **fixest** update.
 
 # insight 0.8.0